youtube-summary/backend/requirements.txt
fastapi
pydantic
youtube_transcript_api
python-dotenv
openai
uvicorn
yt-dlp
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart 